fix: Fix case 30 grid
