+++ packages/topology_optimizer_pkg/tests/conftest.py
@@ -16,6 +16,7 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 import pandera
+import pypowsybl
 import pytest
 from beartype.typing import Generator, Literal, Union
 from confluent_kafka import Consumer, Producer
@@ -24,6 +25,7 @@ from docker.models.containers import Container
 from fsspec.implementations.dirfs import DirFileSystem
 from jaxtyping import Int
 from omegaconf import DictConfig
+from pypowsybl.network import Network
 from sqlmodel import Session, select
 from toop_engine_contingency_analysis.ac_loadflow_service.kafka_client import LongRunningKafkaConsumer
 from toop_engine_dc_solver.example_grids import (
@@ -33,12 +35,15 @@ from toop_engine_dc_solver.example_grids import (
     case57_non_converging,
     complex_grid_battery_hvdc_svc_3w_trafo_data_folder,
     oberrhein_data,
+    three_node_pst_example_folder_powsybl,
 )
-from toop_engine_dc_solver.jax.types import ActionSet
+from toop_engine_dc_solver.jax.types import ActionSet, StaticInformation
 from toop_engine_dc_solver.preprocess import load_grid
+from toop_engine_dc_solver.preprocess.network_data import NetworkData
 from toop_engine_grid_helpers.powsybl.loadflow_parameters import DISTRIBUTED_SLACK
 from toop_engine_grid_helpers.powsybl.powsybl_helpers import save_lf_params_to_fs
 from toop_engine_interfaces.folder_structure import PREPROCESSING_PATHS
+from toop_engine_interfaces.messages.preprocess.preprocess_results import StaticInformationStats
 from toop_engine_interfaces.nminus1_definition import Nminus1Definition, load_nminus1_definition
 from toop_engine_topology_optimizer.ac.storage import ACOptimTopology, create_session
 from toop_engine_topology_optimizer.interfaces.messages.commons import Framework, GridFile, OptimizerType
@@ -492,6 +497,20 @@ def contingency_ids_case_57(n_minus1_definitions_case_57: list[Nminus1Definition
     for cont in contingencies:
         cont_ids.append(cont.id)
     return cont_ids
+
+
+@pytest.fixture
+def create_3_node_pst_example_grid(
+    tmp_path_factory: pytest.TempPathFactory,
+) -> tuple[StaticInformationStats, StaticInformation, NetworkData, Network]:
+    """Create a temporary folder with the three node PST example grid and load it into memory."""
+    tmp_path = tmp_path_factory.mktemp("three_node_pst_example_grid")
+
+    three_node_pst_example_folder_powsybl(tmp_path)
+    filesystem_dir = DirFileSystem(str(tmp_path))
+    stats, static_information, network_data = load_grid(filesystem_dir, pandapower=False)
+    net = pypowsybl.network.load(tmp_path / PREPROCESSING_PATHS["grid_file_path_powsybl"])
+    return stats, static_information, network_data, net
 
 
 @pytest.fixture
